Add Whitelisted property in the Device interface
This patch allows users to manage LE devices in the white list.

If Whitelisted property is set to true the address of the device
is added to the adapter's white list. This will enable the adapter to
automatic connect to this device.

If set to false the address of the device wil be removed from the
white list.

At the moment only public address is supported. Persistent storage is
not implemented yet.