device: Disable unnecessary auto connections
When a remote device is disconnected from some reason and BlueZ attempts
to connect to the remote device again when auto connection is enabled,
the kernel exposes an errorneous number, which is the ECONNABORTED, that
means the remote device connection has been terminated by user and the
host must *NOT* to connect to it again but BlueZ _always_ enable auto
connection as it should not.

This patches checks for the ECONNABORTED in order to avoid enabling auto
connections where connection won't be possible from host to remote device
anymore.