core: Disable connection after local disconnection
This patch disables Bluetooth Low Energy re-connection after a local
initiated disconnection. ECONNABORTED is raised when the adapter is
powered off or when the user calls any disconnect method, for both
cases the re-connection will be triggered again when the adapter is
powered up or if the user calls device Connect method.