gatt: Add decoding Read/Write security
This patch adds the initial handling in the Attribute Server API of
characteristics permission requirements for read and write operations.
Low(no encryption and no authentication) is the default value, if an
external application registering a GATT service wants different
requirements, it should inform using the properties ReadSecurity and
WriteSecurity.